pyproject.toml
[project]
name = "dtcc-less-goo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "faiss-cpu>=1.11.0",
    "fastapi>=0.115.12",
    "openai>=1.81.0",
    "pandas>=2.2.3",
    "pdfminer-six>=20250506",
    "pgvector>=0.4.1",
    "psycopg2-binary>=2.9.10",
    "psycopg[binary]>=3.2.9",
    "pydantic-ai>=0.2.6",
    "pymupdf>=1.26.0",
    "python-dotenv>=1.1.0",
    "sentence-transformers>=4.1.0",
    "sqlalchemy>=2.0.41",
    "tiktoken>=0.9.0",
    "unstructured[pdf]>=0.17.2",
    "uvicorn>=0.34.2",
]
